# docs/04-api/endpoints/openapi-spec.yaml
openapi: 3.1.0
info:
  title: ToolboxAI Roblox Environment
  description: >-
    AI-Powered Educational Roblox Environment - Generate immersive educational
    content with AI agents
  version: 1.0.0
paths:
...
  /lessons:
    get:
      tags:
        - Lessons
      summary: Get Lessons
      description: Get lessons based on user role and filters.
      operationId: get_lessons_lessons_get
      security:
        - HTTPBearer: []
      parameters:
        - name: class_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: Class Id
        - name: subject
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Subject
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            default: 20
            title: Limit
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Get Lessons Lessons Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /lessons/{lesson_id}:
    get:
      tags:
        - Lessons
      summary: Get Lesson Details
      description: Get detailed information about a specific lesson.
      operationId: get_lesson_details_lessons__lesson_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: lesson_id
          in: path
          required: true
          schema:
            type: integer
            title: Lesson Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Lesson Details Lessons  Lesson Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /lessons/:
    post:
      tags:
        - Lessons
      summary: Create Lesson
      description: Create a new lesson (teachers and admins only).
      operationId: create_lesson_lessons__post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Lesson Data
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Lesson Lessons  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
  /lessons/{lesson_id}/progress:
    put:
      tags:
        - Lessons
      summary: Update Lesson Progress
      description: Update student's progress on a lesson.
      operationId: update_lesson_progress_lessons__lesson_id__progress_put
      security:
        - HTTPBearer: []
      parameters:
        - name: lesson_id
          in: path
          required: true
          schema:
            type: integer
            title: Lesson Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Progress Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: >-
                  Response Update Lesson Progress Lessons  Lesson Id  Progress
                  Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /lessons/{lesson_id}/statistics:
    get:
      tags:
        - Lessons
      summary: Get Lesson Statistics
      description: Get statistics for a lesson (teachers and admins only).
      operationId: get_lesson_statistics_lessons__lesson_id__statistics_get
      security:
        - HTTPBearer: []
      parameters:
        - name: lesson_id
          in: path
          required: true
          schema:
            type: integer
            title: Lesson Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: >-
                  Response Get Lesson Statistics Lessons  Lesson Id  Statistics
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer